requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=0.21.1
aiohttp>=3.8.4
boto3>=1.26.137
pydantic>=1.10.7 